Move Deletion into concern
