+++ spec/node/deletion_spec.rb
@@ -1,0 +1,3 @@
+describe Flutterby::Deletion do
+  pending
+end
